requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
openai==1.12.0
requests==2.31.0
pydantic==2.6.1
pydantic-settings==2.1.0
twilio==8.12.0
python-multipart==0.0.9 